backend/auth_service/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=1.10.7
pydantic-extra-types==2.0.0
email-validator==2.0.0.post2
python-jose>=3.3.0
python-multipart>=0.0.6
passlib>=1.7.4
bcrypt>=4.0.1
sqlalchemy>=2.0.9
alembic>=1.10.3
psycopg2-binary>=2.9.6
pyyaml==6.0.1
jsonschema==4.18.6
pyjwt>=2.6.0
redis>=4.5.4
python-dotenv>=1.0.0
asyncpg>=0.27.0
fastapi-limiter==0.1.5
aioredis==2.0.1

# Testing
pytest==7.4.0
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.24.1 